multi_agent/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
langchain>=0.1.16
langchain-core>=0.1.50
langchain-openai>=0.0.5
openai>=1.14.0
python-dotenv>=1.0.1